BufferedAddLite and BufferedDeleteLite
