+++ examples/7.5.1-plugin-bufferedadd.php
@@ -8,10 +8,11 @@ htmlHeader();
 
 // create a client instance and autoload the buffered add plugin
 $client = new Solarium\Client($adapter, $eventDispatcher, $config);
-$buffer = $client->getPlugin('bufferedadd');
+$buffer = $client->getPlugin('bufferedadd'); // or 'bufferedaddlite'
 $buffer->setBufferSize(10); // this is quite low, in most cases you can use a much higher value
 
 // also register an event hook to display what is happening
+// this only works with 'bufferedadd', 'bufferedaddlite' doesn't trigger events
 $client->getEventDispatcher()->addListener(
     Events::PRE_FLUSH,
     function (PreFlushEvent $event) {
