+++ examples/7.5.2-plugin-buffereddelete.php
@@ -8,10 +8,11 @@ htmlHeader();
 
 // create a client instance and autoload the buffered delete plugin
 $client = new Solarium\Client($adapter, $eventDispatcher, $config);
-$buffer = $client->getPlugin('buffereddelete');
+$buffer = $client->getPlugin('buffereddelete'); // or 'buffereddeletelite'
 $buffer->setBufferSize(10); // this is quite low, in most cases you can use a much higher value
 
 // also register an event hook to display what is happening
+// this only works with 'buffereddelete', 'buffereddeletelite' doesn't trigger events
 $client->getEventDispatcher()->addListener(
     Events::PRE_FLUSH,
     function (PreFlushEvent $event) {
