+++ src/Plugin/BufferedAdd/BufferedAddLite.php
@@ -10,24 +10,91 @@
 namespace Solarium\Plugin\BufferedAdd;
 
 use Solarium\Core\Query\DocumentInterface;
-use Solarium\Plugin\BufferedAdd\Event\AddDocument as AddDocumentEvent;
-use Solarium\Plugin\BufferedAdd\Event\PostCommit as PostCommitEvent;
-use Solarium\Plugin\BufferedAdd\Event\PostFlush as PostFlushEvent;
-use Solarium\Plugin\BufferedAdd\Event\PreCommit as PreCommitEvent;
-use Solarium\Plugin\BufferedAdd\Event\PreFlush as PreFlushEvent;
+use Solarium\Plugin\AbstractBufferedUpdate\AbstractBufferedUpdate;
 use Solarium\QueryType\Update\Query\Command\Add as AddCommand;
 use Solarium\QueryType\Update\Result as UpdateResult;
 
 /**
- * Buffered add plugin.
+ * Buffered add lite plugin.
  *
  * If you need to add (or update) a big number of documents to Solr it's much more efficient to do so in batches.
  * This plugin makes this as easy as possible.
  *
- * You can use the lightweight {@see BufferedAddLite} if you don't need the plugin to dispatch events.
+ * Unlike {@see BufferedAdd}, this plugin doesn't dispatch any events.
  */
-class BufferedAdd extends BufferedAddLite
+class BufferedAddLite extends AbstractBufferedUpdate
 {
+    /**
+     * Buffered documents to add.
+     *
+     * @var DocumentInterface[]
+     */
+    protected $buffer = [];
+
+    /**
+     * Set commitWithin time option.
+     *
+     * @param int $time
+     *
+     * @return self Provides fluent interface
+     */
+    public function setCommitWithin(int $time): self
+    {
+        $this->setOption('commitwithin', $time);
+
+        return $this;
+    }
+
+    /**
+     * Get commitWithin time option value.
+     *
+     * @return int|null
+     */
+    public function getCommitWithin(): ?int
+    {
+        return $this->getOption('commitwithin');
+    }
+
+    /**
+     * Set overwrite boolean option.
+     *
+     * @param bool $value
+     *
+     * @return self Provides fluent interface
+     */
+    public function setOverwrite(bool $value): self
+    {
+        $this->setOption('overwrite', $value);
+
+        return $this;
+    }
+
+    /**
+     * Get overwrite boolean option value.
+     *
+     * @return bool|null
+     */
+    public function getOverwrite(): ?bool
+    {
+        return $this->getOption('overwrite');
+    }
+
+    /**
+     * Create a document object instance and add it to the buffer.
+     *
+     * @param array $fields
+     * @param array $boosts
+     *
+     * @return self Provides fluent interface
+     */
+    public function createDocument(array $fields, array $boosts = []): self
+    {
+        $doc = $this->updateQuery->createDocument($fields, $boosts);
+        $this->addDocument($doc);
+
+        return $this;
+    }
+
     /**
      * Add a document.
      *
@@ -39,14 +106,39 @@ class BufferedAdd extends BufferedAddLite
     {
         $this->buffer[] = $document;
 
-        $event = new AddDocumentEvent($document);
-        $this->client->getEventDispatcher()->dispatch($event);
-
         if (\count($this->buffer) === $this->options['buffersize']) {
             $this->flush();
         }
 
         return $this;
+    }
+
+    /**
+     * Add multiple documents.
+     *
+     * @param DocumentInterface[] $documents
+     *
+     * @return self Provides fluent interface
+     */
+    public function addDocuments(array $documents): self
+    {
+        foreach ($documents as $document) {
+            $this->addDocument($document);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Get all documents currently in the buffer.
+     *
+     * Any previously flushed documents will not be included!
+     *
+     * @return DocumentInterface[]
+     */
+    public function getDocuments(): array
+    {
+        return $this->buffer;
     }
 
     /**
@@ -67,26 +159,20 @@ class BufferedAdd extends BufferedAddLite
         $overwrite = $overwrite ?? $this->getOverwrite();
         $commitWithin = $commitWithin ?? $this->getCommitWithin();
 
-        $event = new PreFlushEvent($this->buffer, $overwrite, $commitWithin);
-        $this->client->getEventDispatcher()->dispatch($event);
-
         $command = new AddCommand();
-        $command->setDocuments($event->getBuffer());
+        $command->setDocuments($this->buffer);
 
-        if (null !== $overwrite = $event->getOverwrite()) {
+        if (null !== $overwrite) {
             $command->setOverwrite($overwrite);
         }
 
-        if (null !== $commitWithin = $event->getCommitWithin()) {
+        if (null !== $commitWithin) {
             $command->setCommitWithin($commitWithin);
         }
 
         $this->updateQuery->add(null, $command);
         $result = $this->client->update($this->updateQuery, $this->getEndpoint());
         $this->clear();
-
-        $event = new PostFlushEvent($result);
-        $this->client->getEventDispatcher()->dispatch($event);
 
         return $result;
     }
@@ -107,23 +193,17 @@ class BufferedAdd extends BufferedAddLite
     {
         $overwrite = $overwrite ?? $this->getOverwrite();
 
-        $event = new PreCommitEvent($this->buffer, $overwrite, $softCommit, $waitSearcher, $expungeDeletes);
-        $this->client->getEventDispatcher()->dispatch($event);
-
         $command = new AddCommand();
-        $command->setDocuments($event->getBuffer());
+        $command->setDocuments($this->buffer);
 
-        if (null !== $overwrite = $event->getOverwrite()) {
+        if (null !== $overwrite) {
             $command->setOverwrite($overwrite);
         }
 
         $this->updateQuery->add(null, $command);
-        $this->updateQuery->addCommit($event->getSoftCommit(), $event->getWaitSearcher(), $event->getExpungeDeletes());
+        $this->updateQuery->addCommit($softCommit, $waitSearcher, $expungeDeletes);
         $result = $this->client->update($this->updateQuery, $this->getEndpoint());
         $this->clear();
-
-        $event = new PostCommitEvent($result);
-        $this->client->getEventDispatcher()->dispatch($event);
 
         return $result;
     }
